+++ barapost/src/fastq.py
@@ -35,7 +35,7 @@ def get_read_avg_qual(qual_str):
 # end def get_read_avg_qual
 
 
-def form_packet(fastq_file, packet_size, fmt_func, max_seq_len):
+def form_packet_numseqs(fastq_file, packet_size, fmt_func, max_seq_len):
     """
     Function reads lines from 'fastq_file' and composes a packet of 'packet_size' sequences.
 
@@ -70,34 +70,84 @@ def form_packet(fastq_file, packet_size, fmt_func, max_seq_len):
         qual_dict[read_id[1:]] = avg_qual
     # end for
 
-    if not max_seq_len is None: # prune sequences
+    if max_seq_len < float("inf"): # prune sequences
         packet = prune_seqs(packet, 'l', max_seq_len)
     # end if
 
     return {"fasta": packet, "qual": qual_dict}, eof
+# end def form_packet_numseqs
 
-# end def form_packet
+
+def form_packet_totalbp(fastq_file, packet_size, fmt_func, max_seq_len):
+    """
+    Function reads lines from 'fastq_file' and composes a packet of 'packet_size' base pairs.
+
+    :param fastq_file: file instance from which to read;
+    :type fastq_file: _io.TextIOWrapper or gzip.File;
+    :param packet_size: number of base pairs to retrive from file;
+    :type packet_size: int;
+    :param fmt_func: formating functio nfrom FORMMATING_FUNCS tuple;
+    :param max_seq_len: maximum length of a sequence proessed;
+    :type max_seq_len: int (None if pruning is disabled);
+    """
+
+    packet = ""
+    qual_dict = dict() # {<seq_id>: <read_quality>}
+    eof = False
+
+    totalbp = 0
+
+    while totalbp < packet_size:
+
+        read_id = fmt_func(fastq_file.readline())
+
+        if read_id == "": # if eof is reached, leave now
+            eof = True
+            break
+        # end if
+
+        read_id = fmt_read_id(read_id)
+        seq = fmt_func(fastq_file.readline())
+        fastq_file.readline() # pass comment
+        avg_qual = get_read_avg_qual( fmt_func(fastq_file.readline()) )
+
+        packet += read_id + '\n' + seq + '\n'
+        qual_dict[read_id[1:]] = avg_qual
+
+        totalbp += min(len(seq), max_seq_len)
+    # end for
+
+    if max_seq_len < float("inf"): # prune sequences
+        packet = prune_seqs(packet, 'l', max_seq_len)
+    # end if
+
+    return {"fasta": packet, "qual": qual_dict}, eof
+# end def form_packet_totalbp
 
 
-def fastq_packets(fastq, packet_size, num_done_seqs, saved_packet_size=None, max_seq_len=None):
+def fastq_packets(fastq, packet_size, num_done_seqs, packet_mode=0,
+    saved_packet_size=None, saved_packet_mode=None, max_seq_len=float("inf")):
     """
     Generator yields fasta-formattedpackets of records from fastq file.
     This function passes 'num_done_seqs' sequences (i.e. they will not be processed)
       to 'pass_processed_files'.
 
-    :param fasta: path to fastq file;
-    :type fasta: str;
+    :param fastq: path to fastq file;
+    :type fastq: str;
     :param packet_size: number of sequences to align in one request ('blastn' launching);
     :type packet_size: int;
-    :param reads_at_all: number of sequences in current file;
-    :type reads_at_all: int;
     :param num_done_seqs: number of sequnces in current file that have been already processed;
-    :type num_doce_reads: int;
+    :type num_done_seqs: int;
+    :param packet_mode: packet mode (see -c option);
+    :type packet_mode: int;
     :param saved_packet_size: size of last sent packet from tmp file. Necessary for resumption.
       It will be None, if no tmp file was in classification directory;
     :type saved_packet_size: int;
+    :param saved_packet_mode: mode used whilst formig the last sent packet from tmp file.
+      Necessary for resumption. It will be None, if no tmp file was in classification directory;
+    :type saved_packet_mode: int;
     :param max_seq_len: maximum length of a sequence proessed;
-    :type max_seq_len: int (None if pruning is disabled);
+    :type max_seq_len: int (float("inf") if pruning is disabled);
     """
 
     how_to_open = OPEN_FUNCS[ is_gzipped(fastq) ]
@@ -113,15 +163,28 @@ def fastq_packets(fastq, packet_size, num_done_seqs, saved_packet_size=None, max
         # End of file
         eof = False
 
-        # Here goes check for saved packet size:
+        # Here goes check for saved packet size and mode:
         if not saved_packet_size is None:
             tmp_pack_size = saved_packet_size
         else:
             tmp_pack_size = packet_size
         # end if
 
+        if not saved_packet_mode is None:
+            tmp_pack_mode = saved_packet_mode
+        else:
+            tmp_pack_mode = packet_mode
+        # end if
+
         # Process all remaining sequences with standart packet size:
         while not eof:
+
+            if tmp_pack_mode == 0:
+                form_packet = form_packet_numseqs
+            else:
+                form_packet = form_packet_totalbp
+            # end if
+
             packet, eof = form_packet(fastq_file, tmp_pack_size, fmt_func, max_seq_len)
 
             if packet["fasta"] == "":
@@ -133,6 +196,7 @@ def fastq_packets(fastq, packet_size, num_done_seqs, saved_packet_size=None, max
             # Switch back to standart packet size
             # As Vorotos said, repeated assignment is the best check:
             tmp_pack_size = packet_size
+            tmp_pack_mode = packet_mode
 
             if eof:
                 return